Adding Example Calculation Notebook
This includes an annotated notebook allowing users to fully reproduce the subunit and bond fragmentation threshold calculations for the (T = 3) triangular tiling and its fragment size distributions, as reported by Brunk & Twarock (2021).  The numerical technique was the same for all tilings reported in the paper, with only the stage of generating the polyhedra (and underlying graphs) differing case-by-case, thus - in tandem with Prof Twarock's work on tiling theory - enabling complete reproduction of the reported results.